+++ private-templates-service/.gitignore
@@ -1,0 +1,6 @@
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+dist
+
+# dependencies
+node_modules/
