+++ private-templates-service/package.json
@@ -5,12 +5,16 @@
 	"main": "server.js",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
-    "start:prod": "npm install && node server.js",
-    "start": "node server.js"
+		"start:prod": "npm install && node server.js",
+		"start": "tsc && node dist/server.js"
 	},
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
 		"express": "^4.17.1"
+	},
+	"devDependencies": {
+		"@types/express": "^4.17.2",
+		"typescript": "^3.7.5"
 	}
 }
